1.add okhttp logic; 2.clean code;

+++ downloadlib/src/main/java-gen/com/tony/downloadlib/greendao/DownloadModelDao.java
@@ -25,9 +25,9 @@ public class DownloadModelDao extends AbstractDao<DownloadModel, String> {
     */
     public static class Properties {
         public final static Property Url = new Property(0, String.class, "url", true, "URL");
-        public final static Property DownloadState = new Property(1, int.class, "downloadState", false, "DOWNLOAD_STATE");
-        public final static Property TotalSize = new Property(2, int.class, "totalSize", false, "TOTAL_SIZE");
-        public final static Property DownloadSize = new Property(3, int.class, "downloadSize", false, "DOWNLOAD_SIZE");
+        public final static Property DownloadState = new Property(1, long.class, "downloadState", false, "DOWNLOAD_STATE");
+        public final static Property TotalSize = new Property(2, long.class, "totalSize", false, "TOTAL_SIZE");
+        public final static Property DownloadSize = new Property(3, long.class, "downloadSize", false, "DOWNLOAD_SIZE");
         public final static Property DownloadPath = new Property(4, String.class, "downloadPath", false, "DOWNLOAD_PATH");
     };
 
@@ -102,9 +102,9 @@ public class DownloadModelDao extends AbstractDao<DownloadModel, String> {
     public DownloadModel readEntity(Cursor cursor, int offset) {
         DownloadModel entity = new DownloadModel( //
             cursor.isNull(offset + 0) ? null : cursor.getString(offset + 0), // url
-            cursor.getInt(offset + 1), // downloadState
-            cursor.getInt(offset + 2), // totalSize
-            cursor.getInt(offset + 3), // downloadSize
+            cursor.getLong(offset + 1), // downloadState
+            cursor.getLong(offset + 2), // totalSize
+            cursor.getLong(offset + 3), // downloadSize
             cursor.isNull(offset + 4) ? null : cursor.getString(offset + 4) // downloadPath
         );
         return entity;
@@ -113,9 +113,9 @@ public class DownloadModelDao extends AbstractDao<DownloadModel, String> {
     @Override
     public void readEntity(Cursor cursor, DownloadModel entity, int offset) {
         entity.setUrl(cursor.isNull(offset + 0) ? null : cursor.getString(offset + 0));
-        entity.setDownloadState(cursor.getInt(offset + 1));
-        entity.setTotalSize(cursor.getInt(offset + 2));
-        entity.setDownloadSize(cursor.getInt(offset + 3));
+        entity.setDownloadState(cursor.getLong(offset + 1));
+        entity.setTotalSize(cursor.getLong(offset + 2));
+        entity.setDownloadSize(cursor.getLong(offset + 3));
         entity.setDownloadPath(cursor.isNull(offset + 4) ? null : cursor.getString(offset + 4));
      }
